uuid-api/requirements.txt
property==2.2
mysql-connector-python==8.0.16